news_analyser/scraping/selenium_scraper.py
"""
Selenium-based scraper for JavaScript-heavy websites
"""

import logging
from typing import List, Dict, Optional
from .base_scraper import BaseScraper

logger = logging.getLogger(__name__)


class SeleniumScraper(BaseScraper):
    """
    Scraper using Selenium for JavaScript-rendered content
    """

    def __init__(self, source_name: str, source_url: str, headless: bool = True):
        """
        Initialize Selenium scraper

        Args:
            source_name: Name of the news source
            source_url: Base URL of the website
            headless: Run browser in headless mode
        """
        super().__init__(source_name, source_url)
        self.headless = headless
        self.driver = None

    def _init_driver(self):
        """Initialize Selenium WebDriver"""
        try:
            from selenium import webdriver
            from selenium.webdriver.chrome.options import Options
            from selenium.webdriver.chrome.service import Service
            from webdriver_manager.chrome import ChromeDriverManager

            options = Options()
            if self.headless:
                options.add_argument('--headless')
            options.add_argument('--no-sandbox')
            options.add_argument('--disable-dev-shm-usage')
            options.add_argument('--disable-gpu')
            options.add_argument('user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36')

            service = Service(ChromeDriverManager().install())
            self.driver = webdriver.Chrome(service=service, options=options)
            self.logger.info("Selenium WebDriver initialized")

        except Exception as e:
            self.logger.error(f"Failed to initialize Selenium WebDriver: {e}")
            self.driver = None

    async def scrape(self) -> List[Dict]:
        """
        Scrape articles using Selenium

        Returns:
            List of article dictionaries
        """
        if not self.driver:
            self._init_driver()

        if not self.driver:
            self.logger.error("Selenium driver not available")
            return []

        try:
            import asyncio
            from selenium.webdriver.common.by import By
            from selenium.webdriver.support.ui import WebDriverWait
            from selenium.webdriver.support import expected_conditions as EC

            # Navigate to URL
            await asyncio.to_thread(self.driver.get, self.source_url)

            # Wait for content to load
            wait = WebDriverWait(self.driver, 10)
            await asyncio.to_thread(
                wait.until,
                EC.presence_of_element_located((By.TAG_NAME, "article"))
            )

            # Get page source
            html = self.driver.page_source

            # Parse with BeautifulSoup
            from bs4 import BeautifulSoup
            soup = BeautifulSoup(html, 'lxml')

            # Extract articles (customize based on site structure)
            articles = []

            # This is a placeholder - implement site-specific parsing
            self.logger.warning("Selenium scraping needs site-specific implementation")

            return articles

        except Exception as e:
            self.logger.error(f"Error in Selenium scraping: {e}")
            return []

    async def scrape_article_content(self, url: str) -> Optional[Dict]:
        """
        Scrape article content using Selenium

        Args:
            url: Article URL

        Returns:
            Dictionary with article content
        """
        if not self.driver:
            self._init_driver()

        if not self.driver:
            return None

        try:
            import asyncio

            await asyncio.to_thread(self.driver.get, url)
            await asyncio.sleep(2)  # Wait for dynamic content

            # Extract content (customize based on site)
            content_data = {
                'title': '',
                'content': '',
                'author': '',
                'publish_date': ''
            }

            # Placeholder - implement site-specific extraction

            return content_data

        except Exception as e:
            self.logger.error(f"Error scraping article with Selenium: {e}")
            return None

    def close(self):
        """Close Selenium driver"""
        if self.driver:
            self.driver.quit()
            self.logger.info("Selenium WebDriver closed")

    def __del__(self):
        """Cleanup on deletion"""
        self.close()
