requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
backoff==2.2.1
beautifulsoup4==4.13.3
browser-use==0.1.40
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.6
feedparser==6.0.11
filetype==1.2.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.1
google-api-python-client==2.108.0
google-auth==2.38.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-genai==1.10.0
googleapis-common-protos==1.67.0rc1
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.27
markdownify==0.14.1
monotonic==1.6
oauthlib==3.2.2
ollama==0.4.7
openai==1.72.0
orjson==3.10.16
packaging==24.2
posthog==3.23.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.7
websockets==15.0.1
zstandard==0.23.0
celery==5.4.0
redis==5.0.7
django-celery-results==2.5.1
# ML and NLP dependencies
transformers>=4.36.0
torch>=2.0.0
vaderSentiment>=3.3.2
textblob>=0.17.1
# Scraping dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
aiohttp>=3.9.0
# Financial data
yfinance>=0.2.33
pandas>=2.0.0
numpy>=1.24.0
# API and Web
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
# Caching
django-redis>=5.4.0
hiredis>=2.3.0
# Monitoring and logging
sentry-sdk>=1.40.0
prometheus-client>=0.19.0
# Utilities
python-magic>=0.4.27
pillow>=10.0.0
dateparser>=1.2.0
newspaper3k>=0.2.8
lxml>=4.9.0
